Fix more errorProne
