+++ instrumentation/grizzly-2.3/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/grizzly/GrizzlyIOStrategyTest.java
@@ -7,9 +7,6 @@ package io.opentelemetry.javaagent.instrumentation.grizzly;
 
 import org.glassfish.grizzly.IOStrategy;
 import org.glassfish.grizzly.http.server.NetworkListener;
-import org.glassfish.grizzly.strategies.LeaderFollowerNIOStrategy;
-import org.glassfish.grizzly.strategies.SameThreadIOStrategy;
-import org.glassfish.grizzly.strategies.SimpleDynamicNIOStrategy;
 
 abstract class GrizzlyIOStrategyTest extends GrizzlyTest {
 
@@ -20,28 +17,4 @@ abstract class GrizzlyIOStrategyTest extends GrizzlyTest {
   }
 
   abstract IOStrategy strategy();
-}
-
-class LeaderFollowerTest extends GrizzlyIOStrategyTest {
-
-  @Override
-  IOStrategy strategy() {
-    return LeaderFollowerNIOStrategy.getInstance();
-  }
-}
-
-class SameThreadTest extends GrizzlyIOStrategyTest {
-
-  @Override
-  IOStrategy strategy() {
-    return SameThreadIOStrategy.getInstance();
-  }
-}
-
-class SimpleDynamicTest extends GrizzlyIOStrategyTest {
-
-  @Override
-  IOStrategy strategy() {
-    return SimpleDynamicNIOStrategy.getInstance();
-  }
 }
