+++ instrumentation/grizzly-2.3/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/grizzly/SameThreadTest.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.grizzly;
+
+import org.glassfish.grizzly.IOStrategy;
+import org.glassfish.grizzly.strategies.SameThreadIOStrategy;
+
+class SameThreadTest extends GrizzlyIOStrategyTest {
+
+  @Override
+  IOStrategy strategy() {
+    return SameThreadIOStrategy.getInstance();
+  }
+}
