+++ instrumentation/grizzly-2.3/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/grizzly/SimpleDynamicTest.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.grizzly;
+
+import org.glassfish.grizzly.IOStrategy;
+import org.glassfish.grizzly.strategies.SimpleDynamicNIOStrategy;
+
+class SimpleDynamicTest extends GrizzlyIOStrategyTest {
+
+  @Override
+  IOStrategy strategy() {
+    return SimpleDynamicNIOStrategy.getInstance();
+  }
+}
